fix(Deprecation): split into separate components
FIXES #1323
